Parser: fix error when parsing an "if ... else ..." expression inside of a comprehension
This commit implements a hacky way to avoid errors when parsing an
"if ... else ..." expression within parenthesis in an comprehension's "if".

When parsing a comprehension the "IF_ELSE" flag is disabled to avoid
parsing multiple comprehension's "if" in a row, e.g. "[x for x in l if a if b]",
as an "if ... else ..." expression. But, this causes a side effect, the
following code, "[f for f in l if (a if b is None else c)]", no longer parses without errors.

To solve this the flag "IF_ELSE_WITHIN_PARENS" was created, this enables
the parser to parse the previous code without errors.